backend/notebooks/llm_analysis/requirements.txt
pandas
google-generativeai
python-dotenv
tqdm